Move observe providers to services level
