+++ application/client/src/app/service/session/base.ts
@@ -1,6 +1,7 @@
 import { TabsService, ITabAPI } from '@elements/tabs/service';
+import { Subscriber } from '@platform/env/subscription';
 
-export abstract class Base {
+export abstract class Base extends Subscriber {
     public abstract destroy(): Promise<void>;
     public abstract bind(tab: ITabAPI): void;
     public abstract uuid(): string;
