+++ application/client/src/app/service/session/dependencies/observing/implementations/files.ts
@@ -1,13 +1,10 @@
 import { Provider as Base } from '../provider';
-import { ObserveSource } from '@service/session/dependencies/observe/source';
+import { ObserveSource } from '@service/session/dependencies/observing/source';
 import { IComponentDesc } from '@elements/containers/dynamic/component';
-import { List } from '../../lists/process/component';
+import { List } from '@ui/views/sidebar/observe/lists/file/component';
 import { IMenuItem } from '@ui/service/contextmenu';
-import { opener } from '@service/opener';
-import { Source } from '@platform/types/transport';
-import { State, KEY } from '../../states/process';
 
-export class Provider extends Base<State> {
+export class Provider extends Base {
     private _sources: ObserveSource[] = [];
 
     public contextMenu(source: ObserveSource): IMenuItem[] {
@@ -15,19 +12,10 @@ export class Provider extends Base<State> {
         const observer = source.observer;
         if (observer !== undefined) {
             items.push({
-                caption: 'Stop',
+                caption: 'Stop tailing',
                 handler: () => {
                     observer.abort().catch((err: Error) => {
-                        this.logger.error(`Fail to abort observing (spawning): ${err.message}`);
-                    });
-                },
-            });
-        } else {
-            items.push({
-                caption: 'Repeat Command',
-                handler: () => {
-                    this.repeat(source.source).catch((err: Error) => {
-                        this.logger.error(`Fail to repeat: ${err.message}`);
+                        this.logger.error(`Fail to abort observing (file tailing): ${err.message}`);
                     });
                 },
             });
@@ -36,17 +24,8 @@ export class Provider extends Base<State> {
     }
 
     public update(sources: ObserveSource[]): Provider {
-        this._sources = sources.filter((source) => source.source.is().process);
+        this._sources = sources.filter((source) => source.source.asFile() !== undefined);
         return this;
-    }
-
-    public openNewSessionOptions() {
-        opener
-            .stream(undefined, true, Source.Process)
-            .text()
-            .catch((err: Error) => {
-                this.logger.error(`Fail to open options: ${err.message}`);
-            });
     }
 
     public sources(): ObserveSource[] {
@@ -78,7 +57,7 @@ export class Provider extends Base<State> {
             } => {
                 return {
                     name: (): string => {
-                        return `Commands`;
+                        return `Files`;
                     },
                     desc: (): string => {
                         return this._sources.length > 0 ? this._sources.length.toString() : '';
@@ -130,17 +109,19 @@ export class Provider extends Base<State> {
         };
     }
 
-    public storage(): {
-        get(): State;
-        key(): string;
-    } {
-        return {
-            get: (): State => {
-                return new State();
-            },
-            key: (): string => {
-                return KEY;
-            },
-        };
+    public override getNewSourceError(): Error | undefined {
+        const active = this._sources.find((s) => s.observer !== undefined);
+        if (active !== undefined) {
+            return new Error(`Cannot attach new source while file is tailing`);
+        }
+        const single = this._sources.find(
+            (s) => s.source.is().file && s.source.parent === undefined,
+        );
+        if (single !== undefined) {
+            return new Error(
+                `Single file has been opened. In this case you cannot attach new source(s).`,
+            );
+        }
+        return undefined;
     }
 }
