+++ application/client/src/app/service/session/dependencies/observing/implementations/index.ts
