+++ application/client/src/app/service/session/dependencies/observing/operation.ts
