+++ application/client/src/app/service/session/dependencies/observing/sde.ts
