+++ application/client/src/app/ui/elements/transport/details/serial/component.ts
@@ -3,7 +3,7 @@ import { Ilc, IlcInterface } from '@env/decorators/component';
 import { ChangesDetector } from '@ui/env/extentions/changes';
 import { Session } from '@service/session/session';
 import { SerialTransportSettings } from '@platform/types/transport/serial';
-import { ObserveOperation } from '@service/session/dependencies/observe/operation';
+import { ObserveOperation } from '@service/session/dependencies/observing/operation';
 import {
     Options as AutocompleteOptions,
     AutocompleteInput,
