+++ application/client/src/app/ui/views/sidebar/observe/lists/file/component.ts
@@ -1,13 +1,13 @@
-import { Component, Input, ChangeDetectorRef, ElementRef, AfterContentInit } from '@angular/core';
+import { Component, ChangeDetectorRef, AfterContentInit } from '@angular/core';
 import { Ilc, IlcInterface } from '@env/decorators/component';
-import { ChangesDetector } from '@ui/env/extentions/changes';
-import { Provider } from '../../providers/implementations/files';
 import { Element } from '../../element/element';
 import { File } from '@platform/types/files';
 import { State } from '../../states/files';
 import { IButton } from '../../common/title/component';
 import { components } from '@env/decorators/initial';
 import { Vertical, Horizontal } from '@ui/service/popup';
+import { ListBase } from '../component';
+import { Provider } from '@service/session/dependencies/observing/implementations/files';
 
 @Component({
     selector: 'app-views-observed-list-file',
@@ -15,12 +15,9 @@ import { Vertical, Horizontal } from '@ui/service/popup';
     styleUrls: ['./styles.less'],
 })
 @Ilc()
-export class List extends ChangesDetector implements AfterContentInit {
-    @Input() provider!: Provider;
-
+export class List extends ListBase<State, Provider> implements AfterContentInit {
     public tailing: Element[] = [];
     public offline: Element[] = [];
-    public state!: State;
     public buttons: IButton[] = [
         {
             icon: 'codicon-tasklist',
@@ -56,12 +53,12 @@ export class List extends ChangesDetector implements AfterContentInit {
 
     protected inited: boolean = false;
 
-    constructor(cdRef: ChangeDetectorRef, private _self: ElementRef) {
-        super(cdRef);
+    constructor(cdRef: ChangeDetectorRef) {
+        super(new State(), cdRef);
     }
 
-    public ngAfterContentInit(): void {
-        this.state = this.provider.state;
+    public override ngAfterContentInit(): void {
+        super.ngAfterContentInit();
         this.update();
         this.env().subscriber.register(
             this.provider.subjects.get().updated.subscribe(() => {
