+++ application/client/src/app/ui/views/sidebar/observe/lists/process/component.ts
@@ -1,14 +1,5 @@
-import {
-    Component,
-    Input,
-    ChangeDetectorRef,
-    ElementRef,
-    AfterContentInit,
-    ViewChild,
-} from '@angular/core';
+import { Component, ChangeDetectorRef, AfterContentInit, ViewChild } from '@angular/core';
 import { Ilc, IlcInterface } from '@env/decorators/component';
-import { ChangesDetector } from '@ui/env/extentions/changes';
-import { Provider } from '../../providers/implementations/processes';
 import { Element } from '../../element/element';
 import { Action } from '@ui/tabs/sources/common/actions/action';
 import { QuickSetup } from '../../../../../elements/transport/setup/quick/process/component';
@@ -17,6 +8,8 @@ import { IButton } from '../../common/title/component';
 import { components } from '@env/decorators/initial';
 import { Vertical, Horizontal } from '@ui/service/popup';
 import { State } from '../../states/process';
+import { ListBase } from '../component';
+import { Provider } from '@service/session/dependencies/observing/implementations/processes';
 
 @Component({
     selector: 'app-views-observed-list-process',
@@ -24,9 +17,7 @@ import { State } from '../../states/process';
     styleUrls: ['./styles.less'],
 })
 @Ilc()
-export class List extends ChangesDetector implements AfterContentInit {
-    @Input() provider!: Provider;
-
+export class List extends ListBase<State, Provider> implements AfterContentInit {
     @ViewChild('quicksetupref') public quickSetupRef!: QuickSetup;
 
     public tailing: Element[] = [];
@@ -69,14 +60,13 @@ export class List extends ChangesDetector implements AfterContentInit {
             },
         },
     ];
-    public state!: State;
 
-    constructor(cdRef: ChangeDetectorRef, private _self: ElementRef) {
-        super(cdRef);
+    constructor(cdRef: ChangeDetectorRef) {
+        super(new State(), cdRef);
     }
 
-    public ngAfterContentInit(): void {
-        this.state = this.provider.state;
+    public override ngAfterContentInit(): void {
+        super.ngAfterContentInit();
         this.update();
         this.env().subscriber.register(
             this.provider.subjects.get().updated.subscribe(() => {
