+++ application/client/src/app/ui/views/sidebar/observe/states/files.ts
@@ -4,6 +4,7 @@ import { unique } from '@platform/env/sequence';
 export const KEY: string = unique();
 
 export class State extends Base {
+
     public key(): string {
         return KEY;
     }
