+++ application/client/src/app/ui/views/sidebar/observe/states/process.ts
@@ -4,6 +4,8 @@ import { unique } from '@platform/env/sequence';
 export const KEY: string = unique();
 
 export class State extends Base {
+    public override quicksetup: boolean = true;
+
     public key(): string {
         return KEY;
     }
